% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_helper_simulation_data_extend_TT.R
\name{adjust_vcov_array_country_pd_to_TT}
\alias{adjust_vcov_array_country_pd_to_TT}
\title{Adjust for NAs and extend time series for Vhat matrix}
\usage{
adjust_vcov_array_country_pd_to_TT(vcov_array_country_pd, TT, NN)
}
\arguments{
\item{vcov_array_country_pd}{the true data to adjust and extend}

\item{TT}{time sereis to extend to}

\item{NN}{number of countries}
}
\value{
the adjusted dataset
}
\description{
Adjust for NAs and extend time series for Vhat matrix
}
