% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_helper_simulation_data_extend_TT.R
\name{adjust_w_regs_to_TT}
\alias{adjust_w_regs_to_TT}
\title{Adjust for zeros and extend time series for regressors}
\usage{
adjust_w_regs_to_TT(wregs, TT)
}
\arguments{
\item{wregs}{the true data to adjust and extend}

\item{TT}{time sereis to extend to}
}
\value{
the adjusted dataset
}
\description{
Adjust for zeros and extend time series for regressors
}
