% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_helper_gibbs_sampler_core.R
\name{bdiagByTime}
\alias{bdiagByTime}
\title{Sorts VCOV array by time}
\usage{
bdiagByTime(VhatArray_A, npara, N, TT, Nnpara)
}
\arguments{
\item{VhatArray_A}{VCOV array sorted by cross-section.}

\item{npara}{Anzahl der Parameter (bei uns immer 3)}

\item{N}{cross sectional dimension}

\item{TT}{time series length}

\item{Nnpara}{N * npara}
}
\value{
Sorted VCOV array by time
}
\description{
Sorts VCOV array by time
}
