% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/100_quarto_setup.R
\name{clean_artifacts}
\alias{clean_artifacts}
\title{Clean up artifact directories/files}
\usage{
clean_artifacts(path = getwd())
}
\arguments{
\item{path}{a character string; the default value should be fine}
}
\value{
pure side effect function; returns invisibly
}
\description{
After first creation/renaming via [setup_new_webpage()] there remains the old
artifact directory '.../quartoWebsiteBlueprint'. It arises probably due to
some `RStudio` internal processes and should be deleted. This function does
this safely.
}
