% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/101_quarto_utils.R
\name{clean_include_dir}
\alias{clean_include_dir}
\title{Cleans directory leaving it only with .qmd-files}
\usage{
clean_include_dir(dir_to_check = "./include")
}
\arguments{
\item{dir_to_check}{character giving the relative path to the directory where
the qmd-files are located}
}
\value{
`NULL`; pure side effect function
}
\description{
Cleaning means removing everything but the qmd-files.
}
