% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/101_quarto_utils.R
\name{clean_tmp_dir}
\alias{clean_tmp_dir}
\title{Cleans the tmp-dir directory}
\usage{
clean_tmp_dir(dir_to_check = "./tmp-dir")
}
\arguments{
\item{dir_to_check}{character giving the relative path to the directory where
the qmd-files are located}
}
\value{
`NULL`; pure side effect function
}
\description{
Cleaning means removing everything and re-generating an empty dir
}
