% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_B_post_full_cpp}
\alias{compute_B_post_full_cpp}
\title{Sum in the inverse of the posterior variance for loadings/partial effects}
\usage{
compute_B_post_full_cpp(
  availableObs,
  invOmega0,
  B0_D0,
  fPost,
  w_regs,
  Viarray,
  yiObs,
  selectR
)
}
\arguments{
\item{availableObs}{available observations}

\item{invOmega0}{inverse prior matrix}

\item{fPost}{matrix of factors (backward sampled states via FFBS)}

\item{w_regs}{regressor matrix}

\item{Viarray}{VCOV array (npara x npara x TT) of cross-sectional unit i}

\item{yiObs}{matrix of obersevations}

\item{selectR}{selector matrix for missing y}
}
\value{
a list of two elements, the first being the posterior mean and the
   second the VCM matrix for sampling
}
\description{
Sum in the inverse of the posterior variance for loadings/partial effects
}
