% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_helper_general.R
\name{compute_ig_moments}
\alias{compute_ig_moments}
\title{Compute Mean and Variance of Inverse Gamma Distribution}
\usage{
compute_ig_moments(alpha, beta)
}
\arguments{
\item{alpha}{Shape parameter of the Inverse Gamma distribution}

\item{beta}{Scale parameter of the Inverse Gamma distribution}
}
\value{
A list containing the variance and mean of the Inverse Gamma
}
\description{
Compute Mean and Variance of Inverse Gamma Distribution
}
