% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_helper_general.R
\name{compute_iw_moments}
\alias{compute_iw_moments}
\title{Compute Mean and Variance of Inverse Wishart Distribution}
\usage{
compute_iw_moments(psi, nu, p = 3)
}
\arguments{
\item{psi}{Diagonal elements of the Wishart distribution}

\item{nu}{Degrees of freedom}

\item{p}{Dimension of the Inverse Wishart (default = 3)}
}
\value{
A list containing the variance and mean of the Inverse Wishart
}
\description{
Compute Mean and Variance of Inverse Wishart Distribution
}
