% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_helper_gibbs_sampler_core.R
\name{compute_mat_square_root_V}
\alias{compute_mat_square_root_V}
\title{Matrix square root of V-matrix}
\usage{
compute_mat_square_root_V(v_mat, dim_v, NN_times_TT)
}
\arguments{
\item{v_mat}{V-matrix}

\item{dim_v}{dimension of V-matrix}

\item{NN_times_TT}{number of cross sectional units (`NN`) multiplied by time 
periods (`TT`)}
}
\value{
matrix square root of V
}
\description{
Via helper [mat_sqrt()] using  spectral decomposition
}
