% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/42_helper_measures_computations.R
\name{compute_thin_mcmc}
\alias{compute_thin_mcmc}
\title{Thin Monte Carlo Samples}
\usage{
compute_thin_mcmc(x, thin_freq = NULL)
}
\arguments{
\item{x}{An object of class GibbsOutputIIG, typically the output from a
Monte Carlo sampler. It should contain multidimensional arrays representing
sampled parameter vectors and other relevant matrices.}

\item{thin_freq}{An integer value representing the thinning frequency, i.e.,
the interval at which samples should be retained. For example, a
thin_freq of 10 means every 10th sample is retained.}
}
\value{
Returns a new GibbsOutputIIG object with thinned samples. The
  structure and variables of the input object are preserved, but each
  variable is thinned according to the specified frequency.
}
\description{
This function applies thinning to Monte Carlo samples specifically from a
GibbsOutputIIG class output. Thinning is performed by selecting every nth
sample from the Monte Carlo output, where n is defined by the thinning
frequency. This method can reduce autocorrelation in the samples and decrease
the size of the dataset for analysis.
}
