% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_gibbs_sampler_full_SM_mod02.R
\name{covarianceScale}
\alias{covarianceScale}
\title{Covariance/Corrleation scale}
\usage{
covarianceScale(scale, Var)
}
\arguments{
\item{scale}{Scaling factor}

\item{Var}{VCOV matrix}
}
\value{
returns covariance/correlation scale
}
\description{
Covariance/Corrleation scale
}
