% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/42_plot_measures.R
\name{create_country_plot}
\alias{create_country_plot}
\title{Create Plot for a Given Country}
\usage{
create_country_plot(data_long, country)
}
\arguments{
\item{data_long}{The data in long format.}

\item{country}{The name of the country for which the plot will be created.}
}
\value{
A ggplot object.
}
\description{
Creates a ggplot object for a given country, plotting Gini coefficient values over years.
}
\examples{
data_long <- reshape_data_long(data_raw, c("gini1", "gini_from_gmm"))
create_country_plot(data_long, "Argentina")
}
