% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/42_plot_measures.R
\name{create_me_plots_time_series}
\alias{create_me_plots_time_series}
\title{Generate Time Series Plots for Multiple Regressors}
\usage{
create_me_plots_time_series(
  out_info_KK,
  out_info_TT = NULL,
  grid_x_axis_KK,
  grid_x_axis_TT,
  reg_names,
  RENDER_PLOT = TRUE,
  NO_TITLE = FALSE,
  settings = list(name_measure = "", plot_type = "base", plot_grid = c(3, 5))
)
}
\arguments{
\item{out_info_KK}{A list where each element is a three-dimensional array of
measure values for a specific regressor. The dimensions should correspond
to different entities (e.g., countries), time periods, and measure
statistics (mean values only).}

\item{out_info_TT}{An optional array of dimension `(NN x TT) x KK` with time
related measure infomration to add to the grid-plots}

\item{grid_x_axis_KK}{a four-dimensional array of dimension
\code{NN x TT x GG x KK} where the third dimension gives the number of grid
points and the 4'th dimension the number of regressors \code{KK}}

\item{grid_x_axis_TT}{a 2-dimensional array of dimension
\code{(NNx3) x TT} with appropriate grid-points per parameter}

\item{reg_names}{A vector of character strings specifying the names of the
regressors for which plots will be generated. These names must correspond
to keys in `out_info_KK`.}

\item{RENDER_PLOT}{Logical; if `TRUE` (default), plots are rendered and
displayed. If FALSE, plots are not rendered, allowing for programmatic
control over plot display or further customization.}

\item{NO_TITLE}{Logical; if `FALSE` then default title is generated per
sub-plot; otherwise not}

\item{settings}{A list of settings for plot generation, including:
\itemize{
  \item{\code{name_measure}: }{The measure's name, used for labeling.}
  \item{\code{plot_type}: }{Specifies the plot's type ("base" or "ggplot")
  .}
  \item{\code{plot_grid}: }{A vector specifying the layout of plots in
  terms of rows and columns.}
  \item{\code{WITH_CI}: }{This parameter is ignored in this function as
  confidence intervals are not included in the visualization.}
}}
}
\value{
Implicitly returns a list of plot objects or a composite graphic,
  depending on the `plot_type` and whether `RENDER_PLOT` is TRUE. The
  function is primarily used for its side effect of plotting or preparing
  plots for rendering.
}
\description{
Creates time series plots for each regressor across all entities within the
specified measures. This function facilitates a continuous time series
visualization by integrating time periods into a single plot per entity.
Unlike `create_me_plots_individual`, it focuses on trend visualization over
time without including confidence intervals. The `RENDER_PLOT` flag controls
whether plots are rendered.
}
