% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_helper_data_cleaning.R
\name{data_raw_remove_year_and_save}
\alias{data_raw_remove_year_and_save}
\title{Remove Specific Year Data and Save}
\usage{
data_raw_remove_year_and_save(
  pth_base_data = "./data/input/data-sm",
  data_file = "data_coef_covariates.txt",
  save_file_01 = "data_coef_covariates_2020.rds",
  save_file_02 = "data_coef_covariates_with_ginis_2020.csv",
  year_to_remove = 2021
)
}
\arguments{
\item{pth_base_data}{Base path for data files (default:
"./data/input/data-sm").}

\item{data_file}{Name of the data file (default: "data_coef_covariates.txt").}

\item{year_to_remove}{Year to be removed from the dataset (default: 2021).}

\item{save_file}{Name of the file to save the filtered data (default:
"data_coef_covariates_2020.rds").}
}
\value{
None
}
\description{
This function reads a dataset, filters out a specified year (default 2021),
and saves the filtered dataset as an `.rds` file. It also adds the ginis for
countries and years and saves the data to a `.txt`-file
}
