% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_helper_data_cleaning.R
\name{firstObs_center}
\alias{firstObs_center}
\title{Center values by the first non-NA observation}
\usage{
firstObs_center(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector with values centered around the first non-NA
  observation.
}
\description{
This function centers the values in a vector by subtracting the first non-NA
observation from all elements.
}
