% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_helper_data_cleaning.R
\name{firstObs_center_values}
\alias{firstObs_center_values}
\title{Retrieve the first non-NA observation}
\usage{
firstObs_center_values(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The first non-NA value in the vector.
}
\description{
This function returns the first non-NA observation from a numeric vector.
}
