% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/42_plot_measures.R
\name{generate_country_plots}
\alias{generate_country_plots}
\title{Generate Country-Specific Plots}
\usage{
generate_country_plots(
  pth_data,
  output_measure,
  output_regs_grid = NULL,
  vars_to_use,
  settings,
  name_measure = NULL,
  PLOT = TRUE
)
}
\arguments{
\item{pth_data}{Path to the data file containing relevant data.}

\item{output_measure}{Array of estimated measure coefficient values as
returned via [generate_measures_me()].}

\item{output_regs_grid}{Optional. A grid of regression outputs associated
with measures, to be used for additional data points in the plots.
Default is NULL, indicating no regression grid is used.}

\item{vars_to_use}{Vector of variable names to add as `geom_points` for
comparison with a-posteriori estimation.}

\item{settings}{List of settings for data processing, including:
\itemize{
\item{\code{scale_transform}}{ Numeric value used to scale the measure
values}
\item{\code{y_lab}}{ Label for the y-axis}
\item{\code{x_lab}}{ Label for the x-axis (optional)}
\item{\code{X_TRANSFORMED}}{ Boolean indicating whether to transform the
x-axis values (optional)}
\item{\code{ADD_KI}}{ Boolean indicating whether to add confidence intervals
   to the plots (optional)}
}}

\item{name_measure}{Optional. Character string providing a name for the
measure being plotted. This is used for naming within the plots if
provided. Default is NULL.}

\item{PLOT}{Boolean indicating whether to actually plot the graphs or just
return the data for plotting. Useful for non-interactive environments.
Default is `TRUE`.}
}
\value{
A named list of two elements:
  \itemize{
    \item{\code{data_long}}{ Data in long format suitable for plotting}
    \item{\code{plot_objects}}{ A list of plot objects, one for each country,
       which can be rendered directly if PLOT is TRUE}
  }
}
\description{
This function generates individual plots for each country based on Gini
coefficient data. It processes and reshapes the data, then creates and
arranges plots for each country.
}
