% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_ml_tools.R
\name{generate_f_random_walk}
\alias{generate_f_random_walk}
\title{Simulate random walk factors}
\usage{
generate_f_random_walk(NN_TT, num_y, TT, num_joint_facs)
}
\arguments{
\item{NN_TT}{number of cross sectional units `x` time periods}

\item{num_y}{dimension of the output V-hat matrix (should be smaller then the
current dimension of the `V_sq_adj` as passed via the first argument)}

\item{TT}{number of time periods}

\item{num_joint_facs}{number of joint factors}
}
\value{
simulated random walk factors
}
\description{
Simulate random walk factors
}
