% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/42_full_measures.R
\name{generate_measures_me}
\alias{generate_measures_me}
\title{Computes various a-posterior measures after estimation}
\usage{
generate_measures_me(out_gibbs, transformation_infos, const_num_para = 3)
}
\arguments{
\item{out_gibbs}{the Gibbs output as an object of class `GibbsOutputIIG`}

\item{transformation_infos}{the centering and standardizing vectors for the
`y-measurement` data as a vector for all countries (long index) and all 
parameters (short index) e.g. a vector of length `30` (`3 x 10`, for ten 
countries)}

\item{const_num_para}{numeric value giving the number of parameters in the 
model}
}
\value{
a named list of computed measures
}
\description{
Computes various a-posterior measures after estimation
}
