% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_ml_tools.R
\name{generate_y_simul}
\alias{generate_y_simul}
\title{Simulate measurements/data/observations}
\usage{
generate_y_simul(u, B, f, D = NULL, wRegs = NULL)
}
\arguments{
\item{u}{errors}

\item{B}{loadings}

\item{f}{factors}

\item{D}{marginal regressor effects}

\item{wRegs}{regressors}
}
\value{
simulated data/measurements/observations
}
\description{
Simulate measurements/data/observations
}
