% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_BD_sample}
\alias{get_BD_sample}
\title{Helper to separate joint, idiosyncratic and regressor samples}
\usage{
get_BD_sample(draw, num_jnt_fac, num_y)
}
\arguments{
\item{draw}{full sample draw from truncated normal}

\item{num_jnt_fac}{an integer giving the number of joint factors}

\item{num_y}{integer giving the number of components in the measurements}
}
\value{
a list of thre elements, the first being the idiosyncratic factors,
   the second being the common factor part and the last being the regressors
}
\description{
Helper to separate joint, idiosyncratic and regressor samples
}
