% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/42_plot_measures.R
\name{get_data_measure_plots}
\alias{get_data_measure_plots}
\title{Generate Plots for Data Measures}
\usage{
get_data_measure_plots(
  data_estim,
  output_regs_grid,
  pth_data_gmm,
  vars_to_use,
  names_to_use,
  settings
)
}
\arguments{
\item{data_estim}{An array of estimated measure-coefficient values,
structured similarly to the `out_measures` parameter in
  `data_plot_processing`. This array is used to add new variables to the
  data set based on the estimated measures.}

\item{pth_data_gmm}{The path to the data file, which is used by the
`data_plot_processing` function to read and preprocess the data.}

\item{vars_to_use}{A character vector specifying the variable names to be
included in the long format data. These are the variables that will be
reshaped and used for plotting.}

\item{names_to_use}{A named list or vector specifying the column names to be
created in the data set for each type of measure, used in the
`data_plot_processing` function.}

\item{settings}{A list containing settings for data processing. This should
include at least `scale_transform`, a numeric value used to scale the
measure values in `data_plot_processing`.}

\item{data_regs_grid}{an array of dimension `(NN x num_par) x TT x KK` where
the third  dimension gives the number of regressors (the second time and
the first a is cross section )}
}
\value{
A tibble in long format, suitable for plotting, containing the
  reshaped data with added measure variables.
}
\description{
This function serves as a wrapper for `data_plot_processing` and
`get_data_plot_long`, facilitating the generation of measure-based plots. It
first processes the data by incorporating estimated measures and then
reshapes it into a long format suitable for plotting.
}
