% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_helper_data_cleaning.R
\name{get_data_meta_attributes}
\alias{get_data_meta_attributes}
\title{Group Data Attributes}
\usage{
get_data_meta_attributes(data_gmm)
}
\arguments{
\item{data_gmm}{A data frame containing at least the columns 'year' and
'country'.}
}
\value{
A list containing years, total number of years (TT), countries,
        total number of countries (N), and a name matrix (nameMat).
}
\description{
This function calculates unique years and countries from a dataset and
creates a matrix of country-parameter combinations.
}
