% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/42_plot_measures.R
\name{get_data_mu_long}
\alias{get_data_mu_long}
\title{Reshape Data to Long Format}
\usage{
get_data_mu_long(data_raw, mu_to_use)
}
\arguments{
\item{data_raw}{The raw data as a tibble.}

\item{ginis_to_use}{character giving the name of the `mu`/expectation
variable to use}
}
\value{
A tibble in long format with columns for Gini types and values.
}
\description{
Reshapes the data into a long format suitable for ggplot2, focusing on
specified expectation variable.
}
\examples{
data_raw <- read_and_preprocess_data("/path/to/data", "data.csv", "variable_to_remove")
reshape_data_long(data_raw, c("gini1", "gini_from_gmm"))
}
