% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/42_plot_measures.R
\name{get_data_plot_long}
\alias{get_data_plot_long}
\title{Reshape Data to Long Format}
\usage{
get_data_plot_long(data_raw, var_to_use, nms_to, vals_to)
}
\arguments{
\item{data_raw}{The raw data as a tibble.}

\item{var_to_use}{A character vector of variable names to be used in the
transformation. These variables will be reshaped from wide to long format.}

\item{nms_to}{A string specifying the name of the new column in the long
format that will contain the variable names from `var_to_use`.}

\item{vals_to}{A string specifying the name of the new column in the long
format that will contain the values from the variables in `var_to_use`.}
}
\value{
A tibble in long format with columns as specified by `nms_to` and
  `vals_to`.
}
\description{
Reshapes the data into a long format suitable for ggplot2, focusing on
specified variable(s).
}
\details{
This function transforms the data into a long format, which is particularly
useful for ggplot2 visualizations. It allows for the selection of specific
variables to be reshaped, making it versatile for various types of data beyond
just Gini coefficients.
}
\examples{
\dontrun{
# Example usage with Gini coefficients
data_long_ginis <- get_data_gini_long(data_raw,
                                      ginis_to_use,
                                      "gini_type",
                                      "gini_value")
}
}
