% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_FFBS_helper.R
\name{get_ffbs_missing_obs_adj}
\alias{get_ffbs_missing_obs_adj}
\title{Helper to pre-compute adjustment because of missing values in a KF}
\usage{
get_ffbs_missing_obs_adj(yObs, DwReg, R, C, TT, dimY)
}
\arguments{
\item{yObs}{Y measurements/observations with possible missing values}

\item{DwReg}{DxwReg matrices to be adjusted to the missings in Y}

\item{R}{R matrix to be adjusted to the missings in Y}

\item{C}{C matrix to be adjusted to the missings in Y}

\item{TT}{time series length}

\item{dimY}{length of measurement component}
}
\value{
a list of four elements being lists of length `TT` with adjusted 
   matrices (for C, R, yObs, and `D` times `wReg`)
}
\description{
Helper to pre-compute adjustment because of missing values in a KF
}
