% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_ml_tools.R
\name{get_seq_grid_ml}
\alias{get_seq_grid_ml}
\title{Helper to generate grid around a scalar parameter (centered at true value).}
\usage{
get_seq_grid_ml(
  true_val,
  lower = NULL,
  upper = NULL,
  settings = list(gridlength = 100, grid_width = 0.9)
)
}
\arguments{
\item{true_val}{true parameter value}

\item{lower}{either `NULL` or lower bound}

\item{upper}{either `NULL` or upper bound}

\item{settings}{a named list of two elements, the first giving the number of
grid points as `gridlength = 100` and the second the scaling `grid_width =
0.9` which is used as follows: if both `lower` and `upper` are `NULL`, then
the lower bound is computed as `true_val - true_val * settings$grid_width`
and the upper bound is computed as
`true_val - true_val * settings$grid_width`}
}
\value{
a numeric vector of grid points
}
\description{
Helper to generate grid around a scalar parameter (centered at true value).
}
