% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/42_plot_measures.R
\name{get_single_plot_me}
\alias{get_single_plot_me}
\title{Plot Single Measure with Confidence Intervals, Custom Title, and Type}
\usage{
get_single_plot_me(
  vals_to_plot,
  x_to_plot = NULL,
  pnt_add = NULL,
  settings = list(WITH_CI = FALSE, y_lab = "", x_lab = "", title = NULL, line_col =
    "black", min_max = NULL, type = "", ggplot = FALSE, base_ggplot = NULL)
)
}
\arguments{
\item{vals_to_plot}{A matrix where columns represent the measure values to be
plotted, and, if `WITH_CI` is TRUE, the second and third columns must
contain the upper and lower confidence intervals, respectively.}

\item{x_to_plot}{values for the x-axis; if `NULL` a default index sequence is
is used}

\item{pnt_add}{a vector of x/y coordinates of the point to add}

\item{settings}{A list of settings for the plot, which includes:
\itemize{
  \item{\code{WITH_CI} :}{Logical indicating whether to include confidence
  intervals. If `FALSE`, only measure values are plotted. If `TRUE`, and
  type is "plot", confidence intervals are shown as dashed lines.}
  \item{\code{title} :}{The title of the plot.}
  \item{\code{type} :}{The type of plot ("plot" for a basic plot or "line"
  for a line plot). "line" type cannot include confidence intervals.}
  \item{\code{ggplot} :}{Logical indicating if ggplot2 should be used.
  Defaults to FALSE.}
  \item{\code{y_lab} :}{Y-axis label.}
  \item{\code{x_lab} :}{X-axis label.}
  \item{\code{line_col} :}{Color of the line or points.}
  \item{\code{min_max} :}{Vector specifying the minimum and maximum y-axis
  values.}
  \item{\code{base_ggplot} :}{An existing ggplot object to add the line or
  points to, required if `type` is "line" and `ggplot` is TRUE.}
}}
}
\value{
Generates and potentially returns a plot object based on the
  specified parameters, including measure values and optional confidence
  intervals.
}
\description{
Generates a plot for a single measure, allowing for confidence intervals,
customization of plot type, and visualizing the measure across a specified
dimension. Supports both base R plots and ggplot2, with options for plot
customization.
}
