% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_ml_tools.R
\name{get_u_simul}
\alias{get_u_simul}
\title{Generate errors for Y process simulation}
\usage{
get_u_simul(V_sq_adj, A_means, num_y, TT, NN)
}
\arguments{
\item{V_sq_adj}{3x3 or other dimensional V-hat matrix}

\item{A_means}{}

\item{num_y}{}

\item{TT}{number of time periods}

\item{NN}{number of cross sectional units}
}
\value{
simulated errors
}
\description{
Generate errors for Y process simulation
}
