% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_ml_tools.R
\name{get_V_sq_adj2}
\alias{get_V_sq_adj2}
\title{Helper to compute V-hat adjusted matrix to the dimension of Y.}
\usage{
get_V_sq_adj2(V_sq_adj, NN_TT, num_y, num_v_sq_taken)
}
\arguments{
\item{V_sq_adj}{3x3 or other dimensional V-hat matrix}

\item{NN_TT}{number of cross sectional units `x` time periods}

\item{num_y}{dimension of the output V-hat matrix (should be smaller then the
current dimension of the `V_sq_adj` as passed via the first argument)}
}
\value{
adjusted matrix
}
\description{
Helper to compute V-hat adjusted matrix to the dimension of Y.
}
