% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_helper_general.R
\name{ig_moments}
\alias{ig_moments}
\title{Compute Mean and Variance of Inverse Gamma for Multiple Parameters}
\usage{
ig_moments(alpha_taken, beta_taken)
}
\arguments{
\item{alpha_taken}{Vector of shape parameters of the Inverse Gamma 
distribution}

\item{beta_taken}{Vector of scale parameters}
}
\value{
A list containing matrices of means and variances for each alpha and
   beta
}
\description{
Compute Mean and Variance of Inverse Gamma for Multiple Parameters
}
\examples{
alpha_taken <- c(3, 5, 7, 10)
beta_taken <- c(2, 4, 6, 8)
ig_moments(alpha_taken, beta_taken)
}
