% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_helper_general.R
\name{iw_moments}
\alias{iw_moments}
\title{Compute Mean and Variance of Inverse Wishart for Multiple Parameters}
\usage{
iw_moments(psi_taken, nu_taken, p = 3)
}
\arguments{
\item{psi_taken}{Vector of diagonal elements of the Wishart distribution}

\item{nu_taken}{Vector of degrees of freedom}

\item{p}{Dimension of the Inverse Wishart (default = 3)}
}
\value{
A list containing matrices of means and variances for each psi and nu
}
\description{
Compute Mean and Variance of Inverse Wishart for Multiple Parameters
}
\examples{
psi_taken <- c(7, 17, 27, 57)
nu_taken <- 7:20
iw_moments(psi_taken, nu_taken)
}
