% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_helper_data_cleaning.R
\name{log_adj_VCOV}
\alias{log_adj_VCOV}
\title{Adjust Covariance Matrix with Logarithmic Transformation}
\usage{
log_adj_VCOV(VCOV_array, npara, N, TT, yObs_log)
}
\arguments{
\item{VCOV_array}{Initial 3D array of covariance matrices.}

\item{npara}{Number of parameters.}

\item{N}{Number of countries.}

\item{TT}{Total number of years.}

\item{yObs_log}{Logarithmically transformed observation data.}
}
\value{
A 3D array of adjusted and symmetrized covariance matrices.
}
\description{
This function adjusts a 3D array of covariance matrices using logarithmic 
transformations of the observation data. It symmetrizes the adjusted
matrices.
}
