% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_helper_gibbs_sampler_core.R
\name{makeBi}
\alias{makeBi}
\title{Loadings array}
\usage{
makeBi(npara, N, p, p_joint, B_stack, type = "allidio")
}
\arguments{
\item{npara}{Anzahl der Parameter (bei uns immer 3)}

\item{N}{Cross-sectional dimension (number of countries)}

\item{p}{describe later}

\item{p_joint}{Anzahl gmeinsamer Faktoren (njointfac in FFBS und den helpers)}

\item{B_stack}{describe later}

\item{type}{Modelltyp: Auswahls aus "allidio","countryidio", "countryidio_nomu". Bei uns immer "allidio", d.h. jeder Paramter (jeden Landes) bekommt eigenen idiosynkr. Faktor. Hatte auch mal ausprobiert, dass z.B. nur jedes Land einen bekommt.}
}
\value{
array of loadings
}
\description{
Loadings array
}
