% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_helper_gibbs_sampler_core.R
\name{makeBstart}
\alias{makeBstart}
\title{Loading matrix and array}
\usage{
makeBstart(npara, N, p, p_joint, B_par, type = "allidio")
}
\arguments{
\item{npara}{Anzahl der Parameter (bei uns immer 3)}

\item{N}{Cross-sectional dimension (number of countries)}

\item{p}{describe later}

\item{p_joint}{describe later}

\item{B_par}{describe later}

\item{type}{Modelltyp: Auswahls aus "allidio","countryidio", "countryidio_nomu". Bei uns immer "allidio", d.h. jeder Paramter (jeden Landes) bekommt eigenen idiosynkr. Faktor. Hatte auch mal ausprobiert, dass z.B. nur jedes Land einen bekommt.}
}
\value{
Loading matrix and B for start
}
\description{
Loading matrix and array
}
