% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_helper_gibbs_sampler_core.R
\name{matSqrt}
\alias{matSqrt}
\title{Matrix square root (by spectral decomposition)}
\usage{
matSqrt(V)
}
\arguments{
\item{V}{Normal matrix}
}
\value{
matrix square root of V
}
\description{
Matrix square root (by spectral decomposition)
}
