% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_ml_tools.R
\name{ml_objective_identification_01}
\alias{ml_objective_identification_01}
\title{ML objective function for first identification scheme}
\usage{
ml_objective_identification_01(
  x,
  yObs,
  wReg,
  B_means_used,
  A_means_used,
  Q_used,
  V_sq_adj,
  Phi,
  ccc,
  initX,
  initU,
  initP,
  id_A_mat_optim,
  NN,
  TT,
  num_y,
  NN_TT,
  num_jnt_fac,
  PLOT = FALSE
)
}
\arguments{
\item{x}{argument for parameter vector over which to maximize}

\item{yObs}{observations/measurements/data}

\item{wReg}{regressors}

\item{B_means_used}{true values of loadings}

\item{A_means_used}{true values of adjustment matrix}

\item{Q_used}{true values of transition variances}

\item{V_sq_adj}{VCM matrix of measurements/data/observations}

\item{Phi}{random walk matrix}

\item{ccc}{replacmenet for (non-existent) constant}

\item{initX}{initialization value of X}

\item{initU}{initialization value of U}

\item{initP}{initialization value of P}

\item{id_A_mat_optim}{integer vector giving the id's of the adjustment matrix
over which to maximize}

\item{NN}{number of cross sectional units}

\item{TT}{number of time periods}

\item{num_y}{dimension of the output V-hat matrix (should be smaller then the
current dimension of the `V_sq_adj` as passed via the first argument)}

\item{NN_TT}{number of cross sectional units `x` time periods}
}
\value{
log-likelihood value
}
\description{
This identification scheme fixes the KF-transition variances to 1.
}
