% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_ml_tools.R
\name{ml_run_identification_01}
\alias{ml_run_identification_01}
\title{ML estimation under first identification scheme}
\usage{
ml_run_identification_01(
  true_vals,
  start_vals,
  HESSIAN = FALSE,
  arg_list = arg_list_used
)
}
\arguments{
\item{true_vals}{vector of true parameter values}

\item{start_vals}{starting values for optimization}

\item{HESSIAN}{logical; if `TRUE` then hessian is computed at maxima}

\item{arg_list}{arg list for all arguments (except `x`) of 
[IIGpkg::ml_objective_identification_02]}
}
\value{
a list of two named elements, the results of the optimization as 
  `out_optim` and the print summary `summary_optim`
}
\description{
ML estimation under first identification scheme
}
