% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_gibbs_sampler_B0_Omega0.R
\name{mu_sampler}
\alias{mu_sampler}
\title{MCMC sampler for hyperprior-mean}
\usage{
mu_sampler(
  NN,
  npara,
  nreg,
  njointfac,
  B_i,
  D_i,
  invOmega0,
  mu_b0,
  Sigma_b0,
  selectR,
  type
)
}
\arguments{
\item{NN}{cross sectional units}

\item{npara}{dimension of measurements (of income distribution)}

\item{nreg}{number of regressors}

\item{njointfac}{number of joint factors}

\item{B_i}{B-matrices factorized per i}

\item{D_i}{D-matrices factorized per i}

\item{invOmega0}{inverse of prior matrix}

\item{mu_b0}{hyperprior parameter}

\item{Sigma_b0}{hyperprior parameter}

\item{selectR}{selection matrix}

\item{type}{sampling type}
}
\value{
a single posterior sample for hyperprior-parameter mu
}
\description{
MCMC sampler for hyperprior-mean
}
