% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_class_generators.R
\name{new_GibbsOutputIIG}
\alias{new_GibbsOutputIIG}
\title{Class generator for Gibbs output}
\usage{
new_GibbsOutputIIG(x = vector("list"))
}
\arguments{
\item{x}{output object (a named list) as returned via
[IIGpkg::Gibbs2_SM_SA_sampler] and [IIGpkg::GibbsSSM_2]}
}
\value{
an object of type `GibbsOutputIIG` which is a named list; see the 
   return value of [IIGpkg::GibbsSSM_2] for details.
}
\description{
Class generator for Gibbs output
}
