% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_gibbs_sampler_B0_Omega0.R
\name{omega_sampler}
\alias{omega_sampler}
\title{MCMC sampler for hyperprior-VCM}
\usage{
omega_sampler(NN, B_i, D_i, B0, D0, alpha_b0, beta_b0, selectR)
}
\arguments{
\item{NN}{cross sectional units}

\item{B_i}{B-matrices factorized per i}

\item{D_i}{D-matrices factorized per i}

\item{B0}{B0 samples}

\item{D0}{D0 samples}

\item{alpha_b0}{hyperprior parameter}

\item{beta_b0}{hyperprior parameter}

\item{selectR}{selection matrix}
}
\value{
a single posterior sample for hyperprior-parameter Omega
}
\description{
MCMC sampler for hyperprior-VCM
}
