% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_ml_tools.R
\name{plot_data_ml_simul}
\alias{plot_data_ml_simul}
\title{Helper function to plat}
\usage{
plot_data_ml_simul(yObs, f_simul, num_y, NN, num_joint_facs)
}
\arguments{
\item{yObs}{simulated data/measurements/observations as produced via
[IIGpkg::generate_y_simul]}

\item{f_simul}{simulated factors as produced via 
[IIGpkg::generate_f_random_walk]}

\item{num_y}{dimension of the output V-hat matrix (should be smaller then the
current dimension of the `V_sq_adj` as passed via the first argument)}

\item{NN}{number of cross sectional units}

\item{num_joint_facs}{number of joint factors}
}
\value{
pure side effect function for plotting
}
\description{
Helper function to plat
}
