% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/42_filter_fitted_BF_plots_all.R
\name{plot_factor_diagnostics}
\alias{plot_factor_diagnostics}
\title{Plotting function for fitted vs. true factors.}
\usage{
plot_factor_diagnostics(
  f_true,
  f_pm,
  NN_num_y,
  num_joint_fac,
  settings = list(path_save = ".", plot_grid = c(3, 3))
)
}
\arguments{
\item{f_true}{true posterior factors}

\item{f_pm}{fitted factor (a posteriori) means; possibly after burnin}

\item{NN_num_y}{number of cross sectional units times dimension of the data}

\item{num_joint_fac}{}

\item{settings}{a named list of settings; first element is the path where 
plots are saved and the second is the grid according to which plots are 
arranged i.e. `list(path_save = ".", plot_grid = c(3, 3)`}
}
\value{
pure side effect function returning invisibly `NULL`
}
\description{
Plots posteriori meaned factors vs. their true values.
}
