% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/42_filter_fitted_BF_plots_all.R
\name{plot_fBido_diagnostics}
\alias{plot_fBido_diagnostics}
\title{Plotting function for B*f for idiosyncratic factors}
\usage{
plot_fBido_diagnostics(
  f_true,
  f_pm,
  B_true,
  B_pm,
  NN_num_y,
  settings = list(path_save = ".", plot_grid = c(3, 3))
)
}
\arguments{
\item{f_true}{true posterior factors}

\item{f_pm}{fitted factor (a posteriori) means; possibly after burnin}

\item{B_true}{true loadings}

\item{B_pm}{fitted loading (a posteriori) means; possibly after burnin}

\item{NN_num_y}{number of cross sectional units times dimension of the data}

\item{settings}{a named list of settings; first element is the path where 
plots are saved and the second is the grid according to which plots are 
arranged i.e. `list(path_save = ".", plot_grid = c(3, 3)`}
}
\value{
pure side effect function returning invisibly `NULL`
}
\description{
Idiosyncratic factors: fitted vs. true multiplied by the loadings.
}
