% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/42_filter_fitted_BF_plots_all.R
\name{plot_fit_diagnostics}
\alias{plot_fit_diagnostics}
\title{Plotting function for fitted values and their components.}
\usage{
plot_fit_diagnostics(
  yObs,
  wRegs,
  f_pm,
  B_pm,
  D_pm,
  NN_num_y,
  settings = list(path_save = ".", plot_grid = c(3, 3))
)
}
\arguments{
\item{yObs}{observations (data/measurements)}

\item{wRegs}{regressors}

\item{f_pm}{fitted factor (a posteriori) means; possibly after burnin}

\item{B_pm}{fitted loading (a posteriori) means; possibly after burnin}

\item{D_pm}{fitted regressor marginal (a posteriori) means; possibly after burnin}

\item{NN_num_y}{number of cross sectional units times dimension of the data}

\item{settings}{a named list of settings; first element is the path where 
plots are saved and the second is the grid according to which plots are 
arranged i.e. `list(path_save = ".", plot_grid = c(3, 3)`}
}
\value{
pure side effect function returning invisibly `NULL`
}
\description{
Plots observations (data/measurements), the fit coming from joint factors and
loading, the fit stemming from idiosyncratic factors and loadings, the fit
stemming from regressors (times their marginal effects), and the overall fit
to the data (evaluated at the posterior means).
}
