% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/42_full_measures.R
\name{precompute_measures_me}
\alias{precompute_measures_me}
\title{Precomputes large container that allow to get-posterior measures}
\usage{
precompute_measures_me(out_gibbs, transformation_infos, const_num_para = 3)
}
\arguments{
\item{out_gibbs}{the Gibbs output as an object of class `GibbsOutputIIG`}

\item{transformation_infos}{the centering and standardizing vectors for the
`y-measurement` data as a vector for all countries (long index) and all
parameters (short index) e.g. a vector of length `30` (`3 x 10`, for ten
countries)}

\item{const_num_para}{numeric value giving the number of parameters in the
model}
}
\value{
a named list of precomputed container and regressors/regressor grids
  used to get them
}
\description{
Precomputes large container that allow to get-posterior measures
}
