% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_helper_data_cleaning.R
\name{process_covariance_array}
\alias{process_covariance_array}
\title{Process Covariance Array}
\usage{
process_covariance_array(VCOV_array, npara, N, TT, Diag_VCOV = FALSE)
}
\arguments{
\item{VCOV_array}{Initial 3D array of covariance matrices.}

\item{npara}{Number of parameters (default: 3).}

\item{N}{Number of countries.}

\item{TT}{Total number of years.}

\item{Diag_VCOV}{Boolean flag to convert matrices into diagonal form
(default: FALSE).}
}
\value{
A processed 3D array of covariance matrices.
}
\description{
This function processes a 3D array representing country-period specific
covariance matrices. It symmetrizes the matrices and optionally converts
them into diagonal matrices.
}
