% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_helper_data_cleaning.R
\name{process_observation_data}
\alias{process_observation_data}
\title{Process Observation Data}
\usage{
process_observation_data(data_GMM, npara, TT, countries, years)
}
\arguments{
\item{data_GMM}{A data frame with columns 'a', 'q', 'est_mean', 'country',
and 'year'.}

\item{npara}{Number of parameters (default: 3).}

\item{TT}{Total number of years.}

\item{countries}{Vector of unique countries.}

\item{years}{Vector of unique years.}
}
\value{
A list containing various transformations of the observation data.
}
\description{
This function processes observation data from a dataset, performing log
transformation, centering, standardization, and calculating standard
deviations.
}
