% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_helper_data_cleaning.R
\name{remove_year_and_save}
\alias{remove_year_and_save}
\title{Remove Specific Year Data and Save}
\usage{
remove_year_and_save(
  pth_base_data = "./data/input/data-sm",
  data_file = "data_coef_covariates.txt",
  save_file = "data_coef_covariates_2020.rds",
  year_to_remove = 2021
)
}
\arguments{
\item{pth_base_data}{Base path for data files (default: "./data/input/data-sm").}

\item{data_file}{Name of the data file (default: "data_coef_covariates.txt").}

\item{save_file}{Name of the file to save the filtered data (default: "data_coef_covariates_2020.rds").}

\item{year_to_remove}{Year to be removed from the dataset (default: 2021).}
}
\value{
None
}
\description{
This function reads a dataset, filters out a specified year (default 2021),
and saves the filtered dataset as an RDS file.
}
