% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_B_D_cpp}
\alias{sample_B_D_cpp}
\title{Sample gkh-style (from truncated normal)}
\usage{
sample_B_D_cpp(
  availableObs,
  invOmega0,
  B0_D0,
  fPost,
  w_regs,
  Viarray,
  yiObs,
  selectR,
  lower,
  upper,
  num_jnt_fac,
  num_y
)
}
\arguments{
\item{availableObs}{available observations}

\item{invOmega0}{inverse prior matrix}

\item{fPost}{matrix of factors (backward sampled states via FFBS)}

\item{w_regs}{regressor matrix}

\item{Viarray}{VCOV array (npara x npara x TT) of cross-sectional unit i}

\item{yiObs}{matrix of obersevations}

\item{selectR}{selector matrix for missing y}

\item{invOmega0_B0_D0}{inverse prior matrix multiplied by B0,D0}
}
\value{
a list of two elements, the first being the posterior mean and the 
   second the VCM matrix for sampling
}
\description{
Sample gkh-style (from truncated normal)
}
