% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/42_gibbs_plots_all.R
\name{save_Gibbs_plots}
\alias{save_Gibbs_plots}
\title{Helper for plotting}
\usage{
save_Gibbs_plots(
  Gibbs,
  burnin,
  path,
  nameMat,
  nameReg,
  njointfac,
  type,
  predictionCI = TRUE,
  onlyY = FALSE,
  hierachPrior,
  true_values = NULL
)
}
\arguments{
\item{Gibbs}{Gibbs output object}

\item{burnin}{`integer` giving number of burn in periods}

\item{path}{character giving the path to the where plots are stored}

\item{nameMat}{names of matrices}

\item{nameReg}{names of regressors}

\item{njointfac}{number of joint factors}

\item{type}{e.g. "allidio"}

\item{predictionCI}{logical; if `TRUE`, the prediction intervalls are added}

\item{onlyY}{some setting}

\item{hierachPrior}{if there are hierarchical prior}
}
\value{
side effect function returning the diagnostic plots
}
\description{
Helper for plotting
}
