% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_helper_true_vals.R
\name{set_true_vals}
\alias{set_true_vals}
\title{Generate a list with relevant true (simulation) values}
\usage{
set_true_vals(pth_to_true_vals, ...)
}
\arguments{
\item{pth_to_true_vals}{pth where the `.rds` file is stored}

\item{...}{the relevant values to set, see function body}
}
\value{
invisibly saving an `.rds` file
}
\description{
Generate a list with relevant true (simulation) values
}
\examples{
pth_to_true_vals("./kf_test2.RData")
set_true_vals(pth_to_true_vals,
              VhatArray_A,
              A_means,
              B_means,
              D_means,
              yObs,
              w_regs,
              f_simul)
}
