% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/100_quarto_setup.R
\name{setup_new_webpage}
\alias{setup_new_webpage}
\title{Setup helper for webpage}
\usage{
setup_new_webpage(project_name, path = NULL)
}
\arguments{
\item{project_name}{the new project or package name}

\item{path}{to the blueprint quarto package}
}
\value{
pure side effect function; returns invisibly
}
\description{
A wrapper around [changer::changer()] that changes the package name and `git`
repository of this blueprint. Helps to quickly publish the new webpage
without having to rename all files.
}
