% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_helper_data_cleaning.R
\name{standardize}
\alias{standardize}
\title{Standardize a numeric vector}
\usage{
standardize(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector with each element standardized.
}
\description{
This function standardizes a numeric vector by dividing each element by the 
standard deviation of the vector, ignoring NA values.
}
