% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_helper_data_cleaning.R
\name{standardize_VCOV}
\alias{standardize_VCOV}
\title{Standardize Covariance Matrices}
\usage{
standardize_VCOV(VCOV_array, TT, N, sd_vec_y, npara)
}
\arguments{
\item{VCOV_array}{A 3D array of covariance matrices to be standardized.}

\item{TT}{Total number of years.}

\item{N}{Number of countries.}

\item{sd_vec_y}{Vector of standard deviations for each parameter and country.}

\item{npara}{Number of parameters.}
}
\value{
A 3D array of standardized covariance matrices.
}
\description{
This function standardizes a 3D array of covariance matrices using the
standard deviations of the logarithmically transformed observation data.
}
