% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_helper_data_cleaning.R
\name{standardized_values}
\alias{standardized_values}
\title{Compute the standard deviation of a numeric vector}
\usage{
standardized_values(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The standard deviation of the vector, excluding NA values.
}
\description{
This function calculates the standard deviation of a numeric vector,
excluding NA values.
}
