% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sum_ff_kron_v}
\alias{sum_ff_kron_v}
\title{Sum in the inverse of the posterior variance for loadings/partial effects}
\usage{
sum_ff_kron_v(availableObs, fPost, w_regs, Viarray)
}
\arguments{
\item{availableObs}{available observations}

\item{fPost}{matrix of factors (backward sampled states via FFBS)}

\item{w_regs}{regressor matrix}

\item{Viarray}{VCOV array (npara x npara x TT) of cross-sectional unit i}
}
\value{
summation of kronecker products of appropriate dimension
}
\description{
Sum in the inverse of the posterior variance for loadings/partial effects
}
