% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_helper_gibbs_sampler_core.R
\name{sumffkronV}
\alias{sumffkronV}
\title{Sum in the inverse of the posterior variance for loadings/partial effects}
\usage{
sumffkronV(availableObs, npara, nreg, njointfac, i, fPost, wReg, Viarray, type)
}
\arguments{
\item{availableObs}{available observations}

\item{npara}{number of params (usually 3)}

\item{nreg}{number of regressors}

\item{njointfac}{number of joint factors}

\item{i}{cross-sectional index}

\item{fPost}{backward sampled states (FFBS output)}

\item{wReg}{regressors}

\item{Viarray}{VCOV array (npara x npara x TT) of cross-sectional unit i}

\item{type}{either of "allidio", "countryidio" or "countryidio_nomu"}
}
\value{
cronecker summation part
}
\description{
Sum in the inverse of the posterior variance for loadings/partial effects
}
