% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_helper_gibbs_sampler_core.R
\name{sumfyV}
\alias{sumfyV}
\title{Sum part of posterior mean for vectorized loadings/partial effects}
\usage{
sumfyV(
  availableObs,
  npara,
  nreg,
  njointfac,
  i,
  fPost,
  wReg,
  yiObs,
  Viarray,
  type
)
}
\arguments{
\item{availableObs}{available observations}

\item{npara}{Anzahl der Parameter (bei uns immer 3)}

\item{nreg}{Anzahl der  Regressoren (1 bis 3)}

\item{njointfac}{Anzahl der gemeinsamen Faktoren (zuletzt immer 0)}

\item{i}{cross-sectional index}

\item{fPost}{Kann angegeben werden, wenn FFBS ausgelassen werden soll und man gegeben fester Faktoren samplen möchte. fPost muss dann #Faktoren x T sein. Wenn es fehlt, wird FFBS durchgefuehrt.}

\item{wReg}{Regressoren für die Messgleichung, also die Makros: nreg * N x T (muss natuerlich zu wRegSpec passen)}

\item{yiObs}{Matrix (npara x TT) of cross-sectional unit i}

\item{Viarray}{VCOV array (npara x npara x TT) of cross-sectional unit i}

\item{type}{Modelltyp: Auswahls aus "allidio","countryidio", "countryidio_nomu". Bei uns immer "allidio", d.h. jeder Paramter (jeden Landes) bekommt eigenen idiosynkr. Faktor. Hatte auch mal ausprobiert, dass z.B. nur jedes Land einen bekommt.}
}
\value{
summation part
}
\description{
Sum part of posterior mean for vectorized loadings/partial effects
}
