% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_class_generators.R
\name{summary.GibbsOutputIIG}
\alias{summary.GibbsOutputIIG}
\title{Summary of Gibbs output}
\usage{
\method{summary}{GibbsOutputIIG}(
  x,
  true_vals = NULL,
  burnin = NULL,
  options = list(nrows_out = 9999, sig_digits = 4, roundoff = 2, ci_val = 1.96)
)
}
\arguments{
\item{x}{output object (a named list) as returned via
[IIGpkg::Gibbs2_SM_SA_sampler] and [IIGpkg::GibbsSSM_2]}

\item{true_vals}{a list of object containing the true values}

\item{options}{a list of options:
* `nrows_out`: number of rows to print for summary
* `roundoff`: integer giving the digits to round values in summary
* `ci_val`: value of confidence band; default is 1.96 for 95\%}
}
\value{
pure side effect functions but returns invisibly a list of the 
  different outputs that are printed to the screen
}
\description{
A table with true values (if they exist), Monte Carlo means, confidence
bands, and an indicator if the true value lies in the CI.
}
