% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_helper_gibbs_sampler_core.R
\name{utuSum}
\alias{utuSum}
\title{Sum of squared residuals (u * u') for VCOV/adjustment matrix sampling}
\usage{
utuSum(uSplit, VhatSqrt, TT, N, npara)
}
\arguments{
\item{uSplit}{List (N elements) of resiudal matrices (npara x TT) for each cross-sectional unit}

\item{VhatSqrt}{Array (npara x npara x N*TT) of square-root VCOVs (sorted by cross-sectional unit)}

\item{TT}{time series length}

\item{N}{cross sectional dimension}

\item{npara}{Anzahl der Parameter (bei uns immer 3)}
}
\value{
sum of squared residuals
}
\description{
Sum of squared residuals (u * u') for VCOV/adjustment matrix sampling
}
