% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_B_mean_cpp}
\alias{compute_B_mean_cpp}
\title{Sum in the inverse of the posterior variance for loadings/partial effects}
\usage{
compute_B_mean_cpp(
  Omega,
  invOmega0,
  B0_D0,
  availableObs,
  selectR,
  fPost,
  w_regs,
  yiObs,
  Viarray
)
}
\arguments{
\item{invOmega0}{inverse prior matrix}

\item{availableObs}{available observations}

\item{selectR}{selector matrix for missing y}

\item{fPost}{matrix of factors (backward sampled states via FFBS)}

\item{w_regs}{regressor matrix}

\item{yiObs}{matrix of obersevations}

\item{Viarray}{VCOV array (npara x npara x TT) of cross-sectional unit i}
}
\value{
summation of kronecker products of appropriate dimension
}
\description{
Sum in the inverse of the posterior variance for loadings/partial effects
}
