% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/42_plot_measures.R
\name{create_me_plots_individual}
\alias{create_me_plots_individual}
\title{Generate Individual Measure Plots for Multiple Regressors}
\usage{
create_me_plots_individual(
  out_measures_info_KK,
  reg_names,
  settings = list(name_measure = "", plot_grid = c(4, 5), WITH_CI = TRUE)
)
}
\arguments{
\item{out_measures_info_KK}{A list where each element is a three-dimensional
array of measure values for a specific regressor. The dimensions should
correspond to different entities (e.g., countries), time periods, and
measure statistics (mean, upper, and lower confidence intervals).}

\item{reg_names}{Vector of character strings specifying the names of the
regressors for which plots will be generated. These names must correspond
to keys in `out_measures_info_KK`.}

\item{settings}{List of settings for plot generation, including:
\itemize{
  \item `mfrow`: A vector specifying the layout of plots in terms of rows
  and columns.
  \item `WITH_CI`: Logical indicating whether to include confidence
  intervals in the plots. Defaults to TRUE.
}}
}
\value{
Generates plots for each specified regressor and measure combination
  but does not explicitly return any value.
}
\description{
This function creates individual plots for each regressor within the
specified measures. It iterates over each measure, generating plots for each
point in the measure's dimension (e.g., time series data for multiple
countries). The function supports the inclusion of confidence intervals
based on the settings provided.
}
