% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/42_plot_measures.R
\name{create_me_plots_time_series}
\alias{create_me_plots_time_series}
\title{Generate Time Series Plots for Multiple Regressors}
\usage{
create_me_plots_time_series(
  out_measures_info_KK,
  reg_names,
  settings = list(name_measure = "", mfrow = c(3, 5))
)
}
\arguments{
\item{out_measures_info_KK}{A list where each element is a three-dimensional
array of measure values for a specific regressor. The dimensions should
correspond to different entities (e.g., countries), time periods, and
measure statistics (mean values only, as confidence intervals are not
displayed).}

\item{reg_names}{Vector of character strings specifying the names of the
regressors for which plots will be generated. These names must correspond
to keys in `out_measures_info_KK`.}

\item{settings}{List of settings for plot generation, including:
\itemize{
  \item `mfrow`: A vector specifying the layout of plots in terms of rows
  and columns. This controls the arrangement of time series plots for
  multiple entities.
  \item `WITH_CI`: Logical indicating whether to include confidence
  intervals in the plots. In this function, it is effectively ignored as
  confidence intervals are not supported.
}}
}
\value{
Generates time series plots for each specified regressor and entity
  combination but does not explicitly return any value. The function
  emphasizes trend visualization over time for each entity across all
  specified regressors.
}
\description{
This function creates time series plots for each regressor across all
entities within the specified measures. Unlike `create_me_plots_individual`,
which generates plots for each time period separately, this function
integrates time periods into a single plot per entity, facilitating a
continuous time series visualization. Confidence intervals are not included
in this visualization, focusing instead on the trend over time.
}
