% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/42_plot_measures.R
\name{create_single_country_plot}
\alias{create_single_country_plot}
\title{Create Plot for a Given Country}
\usage{
create_single_country_plot(
  data_long,
  name_country,
  plot_info,
  measure_info,
  estim_infos
)
}
\arguments{
\item{data_long}{The data in long format.}

\item{name_country}{character giving the name of the country for the plot}

\item{plot_info}{A list containing information about the data and plot:
\itemize{
  \item \code{name_country}: character giving the name of the country for
  which the plot will be generated.
  \item \code{y_lab}: label for the y-axis.
}}

\item{measure_info}{A list containing information about the measure points:
\itemize{
  \item \code{vals}: The name(s) of the column(s) in \code{data_long}
  containing the values of measure points.
  \item \code{types}: The name(s) of the column(s) in \code{data_long}
  containing the type information of measure points (for different point
  color).
}}

\item{estim_infos}{A list containing information about the estimations:
\itemize{
  \item \code{mean}: The name of the column in \code{data_long}
  representing the mean value estimate.
  \item \code{ki_upp}: The name of the column in \code{data_long}
  representing the upper confidence band estimate.
  \item \code{ki_low}: The name of the column in \code{data_long}
  representing the lower confidence band estimate.
}}
}
\value{
A ggplot object.
}
\description{
Creates a ggplot object for a given country, plotting specified measure
values over years.
}
