% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/42_plot_measures.R
\name{generate_country_plots}
\alias{generate_country_plots}
\title{Generate Country-Specific Plots}
\usage{
generate_country_plots(
  pth_data,
  output_measure,
  vars_to_use,
  settings,
  name_measure = NULL,
  PLOT = TRUE
)
}
\arguments{
\item{pth_data}{path to the data file containing relevant data.}

\item{output_measure}{Array of estimated measure coefficient values as
returned via [generate_measures_me()].}

\item{vars_to_use}{Vector of variable names to add as `geom_points` for
comparison with a-posteriori estimation.}

\item{settings}{List of settings for data processing, including:alpha
\itemize{
\item{scale_transform}{Numeric value used to scale the measure values}
\item{y_lab}{Label for the y-axis}
}}

\item{names_measures}{character giving the name of the measure; can be left
`NULL` in which case the internal data set is getting a default placeholder
for the variable names}
}
\value{
A named list of three elements:
   \itemize{
      \item{`data_long`}{long format data}
      \item{`plot_objects`}{grid of plots, one for each country}
     }
}
\description{
This function generates individual plots for each country based on Gini
coefficient data. It processes and reshapes the data, then creates and
arranges plots for each country.
}
