% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/42_full_measures.R
\name{generate_measures_me}
\alias{generate_measures_me}
\title{Generates various posterior measures from precomputed results}
\usage{
generate_measures_me(
  par_post_estim,
  post_me,
  regs_to_use,
  transformation_infos,
  const_num_para = 3
)
}
\arguments{
\item{par_post_estim}{estimated posterior fitted values container}

\item{post_me}{estimated posterior fitted values container additionally
generated on a grid and per regressor}

\item{regs_to_use}{a vector of regressors to compute ME measures for}

\item{transformation_infos}{the centering and standardizing vectors for the
`y-measurement` data as a vector for all countries (long index) and all
parameters (short index) e.g. a vector of length `30` (`3 x 10`, for ten
countries)}

\item{const_num_para}{numeric value giving the number of parameters in the
model}
}
\value{
a named list of measures
}
\description{
Precomputed results are obtained from [precompute_measures_me]
}
