% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/42_plot_measures.R
\name{get_single_plot_me}
\alias{get_single_plot_me}
\title{Plot Single Measure with Optional Confidence Intervals}
\usage{
get_single_plot_me(vals_to_plot, WITH_CI = FALSE)
}
\arguments{
\item{vals_to_plot}{Matrix where the first column contains the measure values
to be plotted. If `WITH_CI` is TRUE, the second and third columns must
contain the upper and lower confidence intervals, respectively.}

\item{WITH_CI}{Logical indicating whether to include confidence intervals in
the plot. If FALSE, only the measure values are plotted. If TRUE, dashed
lines represent the confidence intervals.}
}
\value{
A plot is generated but not explicitly returned.
}
\description{
This function generates a plot for a single measure, optionally including
confidence intervals. It is designed to visualize the measure across a
specified dimension, such as time, with the option to include upper and lower
confidence intervals for more detailed analysis.
}
