% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_gibbs_sampler_full_SM_mod02.R
\name{Gibbs2_SM_SA_sampler}
\alias{Gibbs2_SM_SA_sampler}
\title{Title}
\usage{
Gibbs2_SM_SA_sampler(
  p_joint,
  B_par,
  D_par,
  nreg,
  OmegaLoad0Scale,
  OmegaReg0Scale,
  countryA,
  A_diag,
  nu0,
  Psi0,
  shape0,
  rate0,
  yObs,
  wRegSpec,
  wReg,
  Vhat,
  incObsOld = 1e+05,
  incObsNew = 1e+05,
  covScale = 1,
  VhatDiagScale,
  VhatDiagScale_start,
  VdiagEst,
  alpha0,
  beta0,
  N,
  TT,
  storePath,
  itermax,
  scaleA,
  diagA,
  sampleA,
  identification,
  type = "allidio"
)
}
\arguments{
\item{type}{}
}
\description{
Title
}
