% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_helper_gibbs_sampler_core.R
\name{makeDstart}
\alias{makeDstart}
\title{Partial effects matrix and array}
\usage{
makeDstart(npara, N, nreg, D_par)
}
\arguments{
\item{npara}{number of params (usually 3)}

\item{N}{cross sectional dimension}

\item{nreg}{number of regressors}

\item{D_par}{number of D parameters}
}
\value{
partial effects matrix and array for starting
}
\description{
Partial effects matrix and array
}
