% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_helper_gibbs_sampler_core.R
\name{sumffkronV}
\alias{sumffkronV}
\title{Sum in the inverse of the posterior variance for loadings/partial effects}
\usage{
sumffkronV(availableObs, fPost, w_regs, Viarray)
}
\arguments{
\item{availableObs}{available observations}

\item{fPost}{backward sampled states (FFBS output)}

\item{w_regs}{regressor values, typically a matrix}

\item{Viarray}{VCOV array (npara x npara x TT) of cross-sectional unit i}
}
\value{
cronecker summation part
}
\description{
Sum in the inverse of the posterior variance for loadings/partial effects
}
