% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_helper_gibbs_sampler_core.R
\name{sumffkronV}
\alias{sumffkronV}
\title{Sum in the inverse of the posterior variance for loadings/partial effects}
\usage{
sumffkronV(availableObs, fPost, w_regs, Viarray)
}
\arguments{
\item{availableObs}{available observations}

\item{fPost}{backward sampled states (FFBS output)}

\item{Viarray}{VCOV array (npara x npara x TT) of cross-sectional unit i}

\item{id_f}{`integer` (sequence); index for selection of correct factors}

\item{w_reg_info}{a named list of two elements 
\itemize{
   \item{w_reg}{ regressor matrix}
   \item{id_reg}{ `integer` (sequence); index for selection of correct 
   regressors}
}}
}
\value{
cronecker summation part
}
\description{
Sum in the inverse of the posterior variance for loadings/partial effects
}
