% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_helper_gibbs_sampler_core.R
\name{sumfyV}
\alias{sumfyV}
\title{Sum part of posterior mean for vectorized loadings/partial effects}
\usage{
sumfyV(availableObs, i, id_f, fPost, w_reg_info, yiObs, Viarray)
}
\arguments{
\item{availableObs}{available observations}

\item{i}{cross-sectional index}

\item{id_f}{`integer` (sequence); index for selection of correct factors}

\item{fPost}{Kann angegeben werden, wenn FFBS ausgelassen werden soll und man gegeben fester Faktoren samplen möchte. fPost muss dann #Faktoren x T sein. Wenn es fehlt, wird FFBS durchgefuehrt.}

\item{w_reg_info}{a named list of two elements 
\itemize{
   \item{w_reg}{ regressor matrix}
   \item{id_reg}{ `integer` (sequence); index for selection of correct 
   regressors}
}}

\item{yiObs}{Matrix (npara x TT) of cross-sectional unit i}

\item{Viarray}{VCOV array (npara x npara x TT) of cross-sectional unit i}
}
\value{
summation part
}
\description{
Sum part of posterior mean for vectorized loadings/partial effects
}
