% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_helper_gibbs_sampler_core.R
\name{sumfyV}
\alias{sumfyV}
\title{Sum part of posterior mean for vectorized loadings/partial effects}
\usage{
sumfyV(availableObs, fPost, w_regs, yiObs, Viarray)
}
\arguments{
\item{availableObs}{available observations}

\item{fPost}{Kann angegeben werden, wenn FFBS ausgelassen werden soll und man gegeben fester Faktoren samplen möchte. fPost muss dann #Faktoren x T sein. Wenn es fehlt, wird FFBS durchgefuehrt.}

\item{w_regs}{regressor values, typically a matrix}

\item{yiObs}{Matrix (npara x TT) of cross-sectional unit i}

\item{Viarray}{VCOV array (npara x npara x TT) of cross-sectional unit i}
}
\value{
summation part
}
\description{
Sum part of posterior mean for vectorized loadings/partial effects
}
